requirements.txt
kafka-python-ng
openai
flask
python-dotenv
ollama
qdrant-client
